fix: merge FilterWrapper with Filter
