--- src/Molecules/FilterWrapper/README.md
@@ -1,1 +1,0 @@
-## FilterWrapper
