--- src/Molecules/FilterWrapper/index.js
@@ -1,3 +1,0 @@
-import FilterWrapper from './FilterWrapper'
-
-export default FilterWrapper
