+++ src/Molecules/index.js
@@ -23,6 +23,5 @@ export { default as Export } from './Export'
 export { default as List } from './List'
 export { default as EmptyPlaceholder } from './EmptyPlaceholder'
 export { default as Breadcrumb } from './Breadcrumb'
-export { default as FilterWrapper } from './FilterWrapper'
 export { default as Filter } from './Filter'
 export { default as DateRange } from './DateRange'
