+++ src/index.js
@@ -44,6 +44,5 @@ export { EmptyPlaceholder } from './Molecules'
 export { Trans, useTranslation, TransProvider } from './Atoms'
 export { Breadcrumb } from './Molecules'
 export { DateTimePicker } from './Atoms'
-export { FilterWrapper } from './Molecules'
 export { Filter } from './Molecules'
 export { DateRange } from './Molecules'
